week 8 - CardSection.js api통신 추가 및 import문 정리
