+++ src/components/CardSection.js
@@ -1,8 +1,27 @@
+import { useEffect, useState } from "react";
+import { getFolder } from "../api";
 import searchIcon from "../assets/Search.svg";
 import CardList from "./CardList";
 import "./CardSection.css";
 
-export default function CardSection({ cardList }) {
+export default function CardSection() {
+  const [cardList, setCardList] = useState([]);
+
+  useEffect(() => {
+    async function getProFileFolder() {
+      try {
+        const {
+          folder: { links },
+        } = await getFolder();
+        setCardList(links);
+      } catch (error) {
+        console.error(error);
+      }
+    }
+
+    getProFileFolder();
+  }, []);
+
   return (
     <div className="cardSection">
       <div className="cardFrame">
